+++ sdk/formrecognizer/ai-form-recognizer/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Renamed the `FormCountryField` type to `FormCountryRegionField`, and changed the `valueType` discriminant property of that type to `"countryRegion:`.
 - Renamed `ReadingOrder` and `Language` to `FormReadingOrder` and `FormLanguage` to reduce the chance that these types would collide with other types having the same name from other packages.
 - Removed the `KnownStyleName`, `KnownSelectionMarkState`, and `KnownKeyValueType` enums, as they represent simple string enums. The `styleName`, `state`, and `valueType` properties (respectively) now have strong string-enum types.
+- Added the `KnownFormLocale` enum to access the well-known possible values of form locales (the `locale` property of the options parameters for prebuilt model recognition).
 - Migrated to the 2.1 Form Recognizer service endpoint for all REST API calls.
 
 ## 3.1.0-beta.3 (2021-04-06)